Updates OCR scrapper
